Extract objects
- extract objects as separate component
- ToDo: change getter in signals
