+++ src/app/components/profile/objects/objects.component.scss
@@ -1,0 +1,45 @@
+@use '@angular/material' as mat;
+
+.tab-help {
+    * {
+      padding: 0 0.5rem;
+    }
+  }
+
+  .selection {
+    display: flex;
+    background-color: lightgreen;
+  }
+
+  .grid-item {
+    position: relative;
+    cursor: context-menu;
+  
+    .selected {
+      background-color: blue;
+    }
+  
+    .actionbutton {
+      @include mat.elevation(1);
+    }
+  }
+
+  .actionbutton {
+    position: absolute;
+    top: 16px;
+    left: 16px;
+    z-index: 1;
+    background-color: rgba(255, 255, 255, 0.9);
+    display: grid;
+    place-items: center;
+    width: 24px;
+    height: 24px;
+    line-height: 24px;
+    padding: 0;
+    .material-icons {
+      font-size: 16px;
+      width: 16px;
+      height: 16px;
+      line-height: 16px;
+    }
+  }
